clean up and reorg after convo with bitian
